requirements.txt
PyQt5>=5.15.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-hosts>=1.0.3
tensorflow>=2.8.0
pillow>=9.0.0
numpy>=1.21.0
tensorflow-hub>=0.12.0
